# UMEC_healthcare-workforce_1.0.0_swagger.yaml
openapi: 3.0.0
...
servers:
# Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/UMEC/healthcare-workforce/1.0.0
  - description: UMEC Healthcare Workforce API
    url: https://app.swaggerhub.com/UMEC/healthcare-workforce/1.0.0
info:
  description: |
    A series of APIs for the healthcare workforce analytical model and visualization.
  version: "1.0.0"
  title: UMEC Healthcare Workforce API
  license:
    name: Apache 2.0 License
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /user/auth:
    post:
      tags:
        - user
      summary: Set the current user authentication context.
      description: Set the current user into the session.
      operationId: setCurrentUser
      responses:
        default:
          description: successful operation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSessionAuth'
        description: The new user session authentication data.
        required: true        
  /user/session:
    get:
      tags:
        - user
      summary: Get user's session state.
      description: Get the user's session state.
      operationId: getUserSession
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSession'                    
    post:
      tags:
        - user
      summary: Set user's session state.
      description: Set the user's session state.
      operationId: setUserSession
      responses:
        default:
          description: successful operation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSessionSet'
        description: The new session data.
        required: true        
...
components:
  schemas:
...
    UserSession:
      type: object
      properties:
        user_email:
          type: string
          example: email@example.com
        state:
          type: string
          example: json_string        
    UserSessionSet:
      type: object
      properties:
        state:
          type: string
          example: json_string
    UserSessionAuth:
      type: object
      properties:
        user_email:
          type: string
          example: email@example.com
...